python/langgraph/agents/basic-RAG/requirements.txt
langgraph
langgraph-cli
langgraph-sdk
langgraph-checkpoint-sqlite
langchain-community
langchain-core
langchain-openai
python-dotenv
langchain-chroma
langgraph-cli[inmem]
langchain-anthropic
langchain-google-vertexai
langchain-aws
arize-phoenix-otel
openinference-instrumentation-langchain